Setting up to refactor host getter module into and endpoint params getter
