+++ src/get-endpoint.js
